+++ CHANGELOG.md
@@ -20,6 +20,7 @@
       - Covenant of Rock
       - In Every House a Stove
       - Joyous Celebration
+      - Passing of Judgment
 
 ### Updated